+++ CHANGELOG.md
@@ -45,3 +45,4 @@
 - Clarified the usage of `enableEncryption` in `MemoryAdapter`, which uses a simple XOR encryption.
 - Refactored `_CacheItemAdapter` to be a generic class, and updated `setupDataCacheX` to register it correctly, resolving code duplication issues.
 - Made the `TypeAdapterRegistry` more extensible by allowing users to register custom adapters without providing a `typeId`. The `typeId` will be automatically generated.
+- Added support for more storage adapters, including `SQLite` and `shared preferences`.